requirement.txt
acres==0.5.0
annotated-types==0.7.0
antlr4-python3-runtime==4.9.3
anyio==4.12.1
asyncio==4.0.0
attrs==25.4.0
Automat==25.4.16
babel==2.17.0
beautifulsoup4==4.14.3
certifi==2026.1.4
cffi==2.0.0
charset-normalizer==3.4.4
ci-info==0.4.0
click==8.3.1
colour==0.1.5
configobj==5.0.9
configparser==7.2.0
constantly==23.10.4
courlan==1.3.2
cryptography==46.0.3
cssselect==1.3.0
dateparser==1.2.2
deepdiff==8.6.1
defusedxml==0.7.1
distro==1.9.0
drawsvg==2.4.1
et_xmlfile==2.0.0
etelemetry==0.3.1
feedfinder2==0.0.4
feedparser==6.0.12
filelock==3.20.2
google-auth==2.47.0
google-genai==1.57.0
greenlet==3.3.0
h11==0.16.0
htmldate==1.9.4
httpcore==1.0.9
httplib2==0.31.2
httpx==0.28.1
hyperlink==21.0.0
idna==3.11
Incremental==24.11.0
iniconfig==2.3.0
itemadapter==0.13.1
itemloaders==1.3.2
jieba3k==0.35.1
jmespath==1.0.1
joblib==1.5.3
jusText==3.0.2
loguru==0.7.3
looseversion==1.3.0
lxml==6.0.2
lxml_html_clean==0.4.3
Markdown==3.10
markdown-it-py==4.0.0
mdurl==0.1.2
mitreattack-python==5.3.0
networkx==3.6.1
newspaper3k==0.2.8
nibabel==5.3.3
nipype==1.10.0
nltk==3.9.2
numpy==2.4.0
openpyxl==3.1.5
orderly-set==5.5.0
packaging==25.0
pandas==2.3.3
parsel==1.10.0
pathlib==1.0.1
pdfminer.six==20251230
pillow==12.1.0
platformdirs==4.5.1
playwright==1.57.0
pluggy==1.6.0
pooch==1.8.2
Protego==0.5.0
prov==2.1.1
puremagic==1.30
pyasn1==0.6.1
pyasn1_modules==0.4.2
pycparser==2.23
pydantic==2.12.5
pydantic_core==2.41.5
PyDispatcher==2.0.7
pydot==4.0.1
pyee==13.0.0
Pygments==2.19.2
PyMuPDF==1.27.1
pyOpenSSL==25.3.0
pyparsing==3.3.2
pypdfium2==5.3.0
pytest==9.0.2
pytest-base-url==2.1.0
pytest-playwright==0.7.2
python-dateutil==2.9.0.post0
python-dotenv==1.2.1
python-slugify==8.0.4
pytz==2025.2
pyxnat==1.6.4
PyYAML==6.0.3
queuelib==1.8.0
rdflib==7.6.0
regex==2025.11.3
requests==2.32.5
requests-file==3.0.1
rich==14.2.0
rsa==4.9.1
scipy==1.17.0
service-identity==24.2.0
setuptools==80.9.0
sgmllib3k==1.0.0
shellingham==1.5.4
simplejson==3.20.2
six==1.17.0
sniffio==1.3.1
soupsieve==2.8.1
stix2==3.0.1
stix2-patterns==2.0.0
tabulate==0.9.0
taxii2-client==2.3.0
tenacity==9.1.2
text-unidecode==1.3
tinysegmenter==0.3
tld==0.13.1
tldextract==5.3.1
tqdm==4.67.1
trafilatura==2.0.0
traits==7.1.0
Twisted==25.5.0
typer==0.21.1
typing-inspection==0.4.2
typing_extensions==4.15.0
tzdata==2025.3
tzlocal==5.3.1
urllib3==2.6.3
w3lib==2.3.1
websockets==15.0.1
wheel==0.45.1
xlsxwriter==3.2.9
zope.interface==8.2
